+++ src/main/java/com/maxtattoo/bean/request/OrderRequest.java
@@ -9,7 +9,7 @@ public class OrderRequest implements GenericRequest {
 
     private Long orderId;
     private Integer sittingNumber;
-    private Double avgSittingCost;
+    private Integer avgSittingCost;
     private Integer orderPrice;
     private Integer alreadyPaid;
     private Integer prepayment;
